refactor testing package and fix yml file
